meeting assistant/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
streamlit>=1.35.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-dotenv>=1.0.0
openai>=1.30.0
